Fix for default shells other than Bash
I use Fish [1] as my default shell and this causes the preview window to
fail for gh commands (Fish uses bits of non-POSIX syntax).

This is because Fzf uses the default shell (set in the `$SHELL`
environment variable) to run the commands passed as arguments to bind.
Since the script is written in Bash, Bash is already a requirement.
Furthermore, the actions on binding are also written with Bash in mind,
so it makes sense in my opinion to instruct Fzf to use this shell.

[1] https://fishshell.com/